remove ast/tool/extract.h and ast/tool/type.h
