+++ src/core/ast/tool/tool.h
@@ -2,7 +2,6 @@
 #pragma once
 #include "extract_config.h"
 #include "eval.h"
-#include "extract.h"
 #include "type.h"
 #include "liner_eq.h"
 #include "stringer.h"
